[addons][vfs] separate "C" and "C++" to different headers
This place the "C" parts to "kodi/c-api/addon-instance/vfs.h" to have
clean for other languages and to confirm a correct ABI between Kodi
and addon used.
