+++ xbmc/addons/kodi-dev-kit/include/kodi/c-api/addon-instance/CMakeLists.txt
@@ -5,6 +5,7 @@ set(HEADERS audio_decoder.h
             peripheral.h
             pvr.h
             screensaver.h
+            vfs.h
             visualization.h)
 
 if(NOT ENABLE_STATIC_LIBS)
